refact(#94): 네비게이션바 수정
