+++ src/pages/Main/Home/Home.tsx
@@ -6,7 +6,7 @@ import API from "../../../axios/BaseUrl";
 import styled from "styled-components";
 import * as CS from "../../../components/Common/CommonStyle";
 import Header from "../../../components/Main/Header/Header";
-import NavigationBar from "../../../components/Main/NavigationBar";
+import NavigationBar from "../../../components/Main/NavigationBar/NavigationBar";
 import MainOtherProfile from "../../../components/Main/MainOtherProfile";
 import HomeBox from "../../../components/Main/HomeBox";
 import Popup from "../../../components/Common/Popup";
